chatserv/requirements.txt
Django==1.5.1
MySQL-python==1.2.5
South==0.8.4
argparse==1.2.1
distribute==0.6.24
django-extensions==1.3.3
djangorestframework==2.3.13
six==1.6.1
wsgiref==0.1.2
